+++ cmd/sync.go
@@ -120,6 +120,14 @@ func syncCharts(charts []config.ChartConfig, args ...string) error {
 				cmdArgs = append(cmdArgs, chart.Version)
 			}
 		} else {
+			
+			// If the release does not exist do not attempt to delete the release
+			exists := ReleaseExists(chart.Namespace, chart.Release)
+			if ! exists {
+				log.Infof("Skipping '%s/%s' as the release does not exist.", chart.Namespace, chart.Release)
+				continue
+			}
+
 			if IsHelm2() {
 				cmdArgs = append(cmdArgs, "delete")
 				cmdArgs = append(cmdArgs, "--purge")